Implemented Sales api

+++ backend/backend/web/api/sale/schema/sale_base.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel
 
 
 class SaleBase(BaseModel):
-    amount: Optional[float] = None
+    amount: Optional[int] = None
